added jwt & sign up/in & post logic

+++ app/controllers/v1/posts_controller.rb
@@ -1,6 +1,7 @@
 class V1::PostsController < ApplicationController
+    before_action :authorized, only: [:create, :update, :delete]
     def all
-        render json: Post.all.limit(params[:limit].to_i || 25).offset(params[:offset].to_i || 0), status: :ok
+        render json: Post.all.limit(params[:limit] || 25).offset(params[:offset] || 0), status: :ok
     end
 
     def get_post
@@ -8,19 +9,21 @@ class V1::PostsController < ApplicationController
     end
 
     def create
-        post = Post.new(params.require(:userId).permit(:title, :thumb, :content), user_id: params[:userId])
+        post = Post.new(params.permit(:title, :thumb, :content).merge({user_id: @user[:id]}))
         assert_save post
-        render json: post, status: :create
+        render json: post, status: :created
     end
 
     def delete
         post = Post.find(params[:postId])
+        raise Err::Exceptions::DeletePostOtherUser unless post[:user_id] == @user[:id]
         post.destroy
         render json: { success: true }, status: :ok
     end
-
+    
     def update
         post = Post.find(params[:postId])
+        raise Err::Exceptions::UpdPostOtherUser unless post[:user_id] == @user[:id]
         edit_object(post)
         render json: post, status: :ok
     end
